Small changes in the physics
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@805 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/vorticon/ai/sndwave.cpp
@@ -49,7 +49,7 @@ void CObjectAI::sndwave_ai(CObject& object, bool hardmode)
 	{
 		object.sprite = SNDWAVE_RIGHT_FRAME + object.ai.ray.animframe;
 
-		if (x > mp_Map->m_width)
+		if (x > static_cast<int>(mp_Map->m_width) )
 			deleteObj(object);
 		else
 		{
